requirements.txt
streamlit
plotly
psycopg2-binary
python-telegram-bot
